+++ prowler/CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to the **Prowler SDK** are documented in this file.
 - Add `repository_codeowners_change_requires_approval` check for GitHub provider. [(#7753)](https://github.com/prowler-cloud/prowler/pull/7753)
 - Add `repository_sensitive_data_scanner_enabled` check for GitHub provider. [(#7759)](https://github.com/prowler-cloud/prowler/pull/7759)
 - Add `repository_default_branch_requires_signed_commits` check for GitHub provider. [(#7777)](https://github.com/prowler-cloud/prowler/pull/7777)
+- Add `repository_inactive_not_archived` check for GitHub provider. [(#xxxx)](https://github.com/prowler-cloud/prowler/pull/xxxx)
 
 ### Fixed
 - Update CIS 4.0 for M365 provider. [(#7699)](https://github.com/prowler-cloud/prowler/pull/7699)